user scaffold: change password, view change email
On the edit page empty password means password is not set. This means
that admin can create a user without a password. That user cannot log
in, but can ask for password change token to his email, and then
set the password.

That is pretty neat.

In addition, 'new user' button is now on top, as list might be long.

Signed-off-by: Zeljko <zeljko@zwr.fi>