competition format
This could have been a constant, but will make it easier
to setup associations this way. There is currently only
one CF, and there will be only several in the beginning.

Each will apply the flow and will have to have some
corresponding code. For instance, in the case of 'two
races combined', controller for the competition should show
standings (and sort) based on the combined times of both races.

Signed-off-by: Zeljko <zeljko@zwr.fi>