create competition scaffold
Generate scaffold for competition, that belongs to an Event
(opposite has_many). Competition also belongs to comp. format.
It has a status, which is a constant from array
Compettition::Status.

Competitions are managed by MasterController, which means that
only a master can list them (not in this commit), while others
can view them if they get the index.

Competitions are shown with an event.

Signed-off-by: Zeljko <zeljko@zwr.fi>