migration: competitors:user must not be null
This is to make sure seeds are OK, as you cannot make an application
without setting a valid user.

This might actually change later as we add redundancy to the
competitor, covering, for instance, a case where user object
is deleted (i.e. user deletes her profile) and a competitor in
a competition stays.

And it already showed us problems with the user with three
names and several others. This commit fixes them all, hopefully.

Signed-off-by: Zeljko <zeljko@zwr.fi>