sort and format results
Results are now sorted by winner (or by number if no
runs are marked yet), and times are displayed in some
sort of format.

Note: we had to load eagerly event and all sub-objects as
calculations are going up and down the hierarchy. We have also
made some 'remembering' of the calculated values.

The queries are still painfully slow, and it is clear that some
redundancy is needed for this system to go into production.

Signed-off-by: Zeljko <zeljko@zwr.fi>